Do not place unused 0th bin into the database.  In theory this should also be done to the netCDF file, but this has legacy implications.  I will modify SourceNetCDF in aeros to adjust the data accordingly.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@3459 640d5228-2204-0410-8d71-8f46fa6850b4

+++ nimbus/include/psql.h
@@ -213,7 +213,7 @@ protected:
    * transmission to the ground.
    */
   void
-  addVectorToAllStreams(const NR_TYPE *values, int nValues, bool xmitToGround);
+  addVectorToAllStreams(const NR_TYPE *values, size_t nValues, bool xmitToGround);
 
   /**
    * Remove any trailing spaces from the string @p target.
